+++ App.js
@@ -1,6 +1,7 @@
-import * as React from 'react';
+import React, { useEffect } from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import SplashScreen from 'react-native-splash-screen';
 
 import HomeScreen from './app/screens/Home/Index';
 import SettingsScreen from './app/screens/Settings/Index';
@@ -9,6 +10,10 @@ import BookScreen from './app/screens/Book/Index';
 const Tab = createBottomTabNavigator();
 
 const App: () => Node = () => {
+    useEffect(() => {
+        SplashScreen.hide();
+    }, []);
+
     return (
         <NavigationContainer>
             <Tab.Navigator>