+++ android/app/src/main/java/com/appproyectosbeta/MainActivity.java
@@ -3,8 +3,15 @@ package com.appproyectosbeta;
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
 import com.facebook.react.ReactRootView;
+import org.devio.rn.splashscreen.SplashScreen;
+import android.os.Bundle;
 
 public class MainActivity extends ReactActivity {
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+      SplashScreen.show(this,  R.style.SplashTheme, true);
+      super.onCreate(savedInstanceState);
+  }
 
   /**
    * Returns the name of the main component registered from JavaScript. This is used to schedule